Working on compiling fancy to rubinius bytecode.
I first tried to port the rubinius compiler chain to fancy and use
that to compile to bytecode. This path was already being experimented
at the rubinius_vm branch. However I noticed it would take a large
ammount of work to port rubinius compiler to fancy by hand.

So the approach I'm takin is this:
Have fancy output fancy sexp using only ruby literals, so that
we can implement a compiler in ruby using rubinius infrastructure.

My first target is compiling a hello world program that would run
on rubinius.

The compiler is invoked using rubinius:

  rbx ./rubiniuxvm/compiler.rb ./rubiniusvm/hello.fnc

The compiler does the following:

  Use fancy to output the fancy program as a ruby sexp (--rsexp option)
  Evaluate the sexp in ruby, and then create AST objects (in ruby)
  Use the rubinius generator to create bytecode
  Use the rest of the compiler chain (encoder, writer, etc)