Started with reimplementing the rbx/fancy_ext/ stuff in lib/rubinius/, if possible, and removed the ruby versions from rbx/fancy_ext/.
We need to start thinking about how to deal with calling ruby methods. IMHO, it makes more sense to use the ruby-args syntax (method: ~[args] for any kind of ruby call since we shouldn't clutter up or redefine existing ruby methods all over the place.
This does mean that for all methods that take arguments, we'd have to write our own wrapper method but we can easily do that with some nice metaprogramming and helper methods.
The upside is any existing ruby code won't need to go through our wrapper code, which is kind of important.
So basically, my suggestion is using the "~" syntax as a splat-operator in ruby-land.

E.g.:
    obj foo: ~[1,2,3]
=>  obj.foo(1,2,3)

and:
    obj foo: ~[1,2, ~[3,4,5]]
=>  obj.foo(1,2,3,4,5)

Comments?