rbx/parser/test.rb prints a bootstrapped hello world!
Now the test.rb file loads the compiled fancy runtime (compiled in lib/*.fyc, lib/rubinius/*.fyc)
but we had a problem...

In fancy, we define methods that take no arguments like (println was just a case)

  def class Object {
    def println {
       "prints itself to output"
       ...
    }
  }

The problem we were facing is that some ruby libraries (in out case, the rubinius compiler tools)
did expect those methods with those exact names to take parameters, but since we just openned the
ruby class and overwrote the method to no longer take arguments, that was failing.

Methods from fancy that take arguments have no problem because their names contain ':' which is very
unlike ruby method names. (eg, in fancy we have foo:bar method names)

So the best solution I found was to use a different name for fancy methods that don't take arguments.
The current compiler prepends a ':' to methods not taking arguments, so that we don't override
the ruby ones, and ruby libraries can use them as they expect.